refactor(session): extract useSolveChannel from DesignSession
The /ws socket, the latest-wins `_seq` protocol counters, syncSolving,
requestSolve/cancelSolve, the busy-chrome dwell effect and its derived
`stale` move whole into components/session/useSolveChannel.ts, together with
the module-level WS_URL. Every moved body is byte-identical, including the
per-socket cutsSender closure and its identity guard in dropCutsSender, and
both dep arrays ([solving], [active]) are literal moves.

controlsRef stays in DesignSession and is passed in: its useRef initializer
calls buildRequest, so it must be constructed before the channel that reads
it. For the same reason sessionIdRef stays behind — it is read only by
buildRequest, so moving it into the hook would make the hook's own arguments
depend on its return value. buildRequest is therefore unchanged.

The hook is called immediately after controlsRef, which moves its two
effects ahead of the live-metrics, combo-reset, solve, preview and analysis
effects. Verified inert: nothing else touches shownAtRef/showBusy, and the
one write->read pair (solve effect -> requestSolve -> wsRef) early-returns
identically either way, since a socket created earlier in the same commit is
CONNECTING, not OPEN.
