+++ src/antennaknobs/web/frontend/src/components/session/DesignSession.tsx
@@ -63,13 +63,6 @@ import type {
 } from "../../lib/api";
 import { BackendConfigModal } from "../backend/BackendConfigModal";
 import { ParamForm } from "../params/ParamForm";
-import {
-  cutsWsSend,
-  flushCutsWsPending,
-  resolveCutsWsMessage,
-  setCutsWsSend,
-  type CutsWsMessage,
-} from "../charts/cuts";
 import type { PatternData, PatternMetrics } from "../charts/types";
 import {
   ThemeContext,
@@ -98,6 +91,7 @@ import { SolveOverlays } from "./SolveOverlays";
 import { SolverSlotTabs } from "./SolverSlotTabs";
 import { useDesignCatalog } from "./useDesignCatalog";
 import { useMobileCarousel } from "./useMobileCarousel";
+import { useSolveChannel } from "./useSolveChannel";
 import {
   type OptimizeResult,
   type OptObjective,
@@ -111,11 +105,6 @@ import {
 // O(1/N) trajectories clearly. Same ladder across backends so the curves
 // are directly comparable when the user switches slots.
 const CONVERGE_N_VALUES: number[] = [8, 12, 17, 24, 34, 48, 68];
-
-// Match the page's scheme: a wss:// upgrade is required on HTTPS pages (e.g. the
-// deployed site behind Fly's force_https), where browsers block insecure ws://
-// as mixed content. Plain ws:// only works on http:// (local dev).
-const WS_URL = `${window.location.protocol === "https:" ? "wss" : "ws"}://${window.location.host}/ws`;
 
 // One antenna design session: the entire left sidebar + right stage plus all
 // the state, effects, and the WebSocket that drive them. The shell (`App`,
@@ -611,15 +600,6 @@ export function DesignSession({ id, active }: { id: number; active: boolean }) {
     preview?.multi_feed ??
     currentExample?.multi_feed ??
     false;
-  const [status, setStatus] = useState<"connecting" | "open" | "closed">("connecting");
-  const [rttMs, setRttMs] = useState<number | null>(null);
-  // True whenever a main solve is outstanding (in flight or queued) — i.e. the
-  // displayed analysis isn't current yet. `showBusy` is the *debounced* view of
-  // it: the progress bar / panel dimming only appear once a solve outlasts
-  // ~300 ms, so fast updates (cache hits, small designs) snap in cleanly
-  // without a flash of busy chrome.
-  const [solving, setSolving] = useState(false);
-  const [showBusy, setShowBusy] = useState(false);
   const [sweep, setSweep] = useState<SweepData | null>(null);
   const [sweepRunning, setSweepRunning] = useState(false);
   // Smith-chart overlay toggles. Both are debounced sweeps that re-fire
@@ -699,17 +679,6 @@ export function DesignSession({ id, active }: { id: number; active: boolean }) {
     setComboApproved(true);
     setSolverWarning(false);
     setAutoSim(false);
-  }
-  // Cancel an IN-FLIGHT solve: stop waiting and discard its result. The server
-  // keeps computing (its /ws loop is sequential and a running MoM solve can't be
-  // interrupted), so this cancels the wait, not the computation.
-  function cancelSolve() {
-    if (lastSentSeqRef.current <= lastReceivedSeqRef.current) return; // nothing in flight
-    // Mark every seq sent so far as cancelled: onmessage will advance the
-    // received watermark for these but drop their results. A newer knob change
-    // bumps lastSentSeq past this and solves again.
-    canceledThroughSeqRef.current = lastSentSeqRef.current;
-    syncSolving();
   }
 
   // Schema-driven design-freq link: when the active example has any
@@ -805,23 +774,12 @@ export function DesignSession({ id, active }: { id: number; active: boolean }) {
   // this signature actually changed, so it skips the redundant refetch right
   // after an antenna switch (whose preview the switch effect already built).
   const previewSigRef = useRef<string | null>(null);
-  // Timestamp (performance.now) when the busy chrome last became visible, so
-  // the reveal effect can enforce a minimum-visible window. null = not shown.
-  const shownAtRef = useRef<number | null>(null);
   // Latest selected antenna, mirrored into a ref so the (mount-once) WebSocket
   // onmessage handler can drop responses for an antenna the user already
   // switched away from. Updated every render — cheap and always current.
   const geometryRef = useRef(geometry);
   geometryRef.current = geometry;
 
-  const wsRef = useRef<WebSocket | null>(null);
-  // Latest-wins /ws protocol counters. Every knob change is sent eagerly with a
-  // monotonic `_seq`; the server keeps only the freshest queued request and may
-  // skip-send superseded results, so the client orders and prunes by `_seq`. A
-  // solve is outstanding iff more has been sent than received. These live in
-  // refs so they survive StrictMode/HMR socket teardown — the counter must
-  // never rewind below what's already been received.
-  const seqRef = useRef(0); // last _seq assigned (monotonic, never reset)
   // Solve-lane session id (issue #382): one per workbench tab (A/B compare
   // tabs are separate App instances, hence separate sessions). The server
   // keys its single-lane scheduler on this — everything this tab asks for
@@ -831,11 +789,6 @@ export function DesignSession({ id, active }: { id: number; active: boolean }) {
       ? crypto.randomUUID()
       : `s-${Math.random().toString(36).slice(2)}`,
   );
-  const lastSentSeqRef = useRef(0); // highest _seq put on the wire
-  const lastReceivedSeqRef = useRef(0); // highest _seq received or implicitly acked
-  const canceledThroughSeqRef = useRef(0); // drop rendering for _seq <= this
-  const sentAtRef = useRef<Map<number, number>>(new Map()); // _seq → send time (RTT)
-  const solveRafRef = useRef<number | null>(null); // trailing-edge rAF throttle handle
 
   function buildRequest(): SolveRequest {
     // ground_model is shared across backends (εr=10, σ=0.002 for the finite
@@ -1119,6 +1072,28 @@ export function DesignSession({ id, active }: { id: number; active: boolean }) {
   // The latest control values, used to send a new request when the prior one
   // completes (drops intermediate values rather than queuing them all up).
   const controlsRef = useRef<SolveRequest>(buildRequest());
+
+  // The /ws solve channel (#642 seam 5b-3): the socket, the latest-wins `_seq`
+  // protocol and the busy-chrome dwell. Called right after controlsRef — the
+  // channel reads that ref on every send, and the solve effect below reaches
+  // for requestSolve.
+  const {
+    status,
+    rttMs,
+    solving,
+    showBusy,
+    stale,
+    requestSolve,
+    cancelSolve,
+    seqRef,
+  } = useSolveChannel({
+    active,
+    controlsRef,
+    geometryRef,
+    previewSigRef,
+    setResult,
+    setSolveError,
+  });
 
   // --- Pattern compare (pin / ghost overlay) --------------------------------
   // Pin the current pattern: snapshot the solve response (for the ghost trace)
@@ -1785,202 +1760,6 @@ export function DesignSession({ id, active }: { id: number; active: boolean }) {
       if (patternAbortRef.current === controller) patternAbortRef.current = null;
     }
   }
-
-  // Mirror the seq counters into `solving` state so the UI can react. Called
-  // wherever the sent / received / cancel watermarks move. A solve reads as
-  // running when more has been sent than received — unless everything
-  // outstanding was cancelled (lastSentSeq hasn't advanced past the cancel
-  // watermark), in which case the wait is over even though a doomed response
-  // is still coming.
-  function syncSolving() {
-    setSolving(
-      lastSentSeqRef.current > lastReceivedSeqRef.current &&
-        lastSentSeqRef.current > canceledThroughSeqRef.current,
-    );
-  }
-
-  // Busy-chrome reveal with two guards:
-  //  - dwell: only show once a solve has been outstanding >BUSY_DWELL_MS. 1 s
-  //    is the classic "flow of thought" threshold — below it users tolerate the
-  //    wait without feedback; at/above it the bar reassures them it's working.
-  //    A solve that finishes sooner clears the timer in cleanup, so the bar
-  //    never flips on for quick updates.
-  //  - min-visible: once shown, keep it up at least BUSY_MIN_VISIBLE_MS so a
-  //    solve that lands just past the dwell can't make it sub-perceptibly
-  //    flash.
-  const BUSY_DWELL_MS = 1000;
-  const BUSY_MIN_VISIBLE_MS = 400;
-  useEffect(() => {
-    if (solving) {
-      const t = window.setTimeout(() => {
-        shownAtRef.current = performance.now();
-        setShowBusy(true);
-      }, BUSY_DWELL_MS);
-      return () => window.clearTimeout(t);
-    }
-    // Solve finished. If the bar never showed (fast solve), hide immediately;
-    // otherwise hold it for the remainder of the minimum-visible window.
-    if (shownAtRef.current === null) {
-      setShowBusy(false);
-      return;
-    }
-    const remaining =
-      BUSY_MIN_VISIBLE_MS - (performance.now() - shownAtRef.current);
-    if (remaining <= 0) {
-      shownAtRef.current = null;
-      setShowBusy(false);
-      return;
-    }
-    const t = window.setTimeout(() => {
-      shownAtRef.current = null;
-      setShowBusy(false);
-    }, remaining);
-    return () => window.clearTimeout(t);
-  }, [solving]);
-
-  // The progress bar (`showBusy`) honors the min-visible window so it can't
-  // flash, but the *dimming* and the "solving…" label mean "what you're
-  // looking at is stale" — so they must clear the instant the result lands,
-  // even while the bar lingers out its minimum. `solving` flips false
-  // immediately on result-land, so `showBusy && solving` is exactly that: dim
-  // only after the dwell (showBusy) AND while genuinely still solving.
-  const stale = showBusy && solving;
-
-  function requestSolve() {
-    const ws = wsRef.current;
-    if (!ws || ws.readyState !== WebSocket.OPEN) {
-      // Can't send now. onopen resends controlsRef.current on (re)connect, so
-      // the latest state is solved as soon as the socket comes up.
-      return;
-    }
-    // Trailing-edge rAF throttle: coalesce a burst of knob changes within one
-    // animation frame to a single send of the latest controls. Bounds upload to
-    // ≤~60 msg/s during a drag and keeps localhost message churn near what the
-    // old one-in-flight gate produced; the server's latest-wins mailbox squashes
-    // whatever still piles up. The freshest value always wins within the frame.
-    if (solveRafRef.current !== null) return;
-    solveRafRef.current = requestAnimationFrame(() => {
-      solveRafRef.current = null;
-      const sock = wsRef.current;
-      if (!sock || sock.readyState !== WebSocket.OPEN) return;
-      const seq = ++seqRef.current;
-      lastSentSeqRef.current = seq;
-      sentAtRef.current.set(seq, performance.now());
-      sock.send(JSON.stringify({ ...controlsRef.current, _seq: seq }));
-      // Keep the preview signature current so that toggling Live *off* right
-      // after a solve doesn't see a stale signature and needlessly refetch the
-      // wireframe / drop the just-solved result — the solved geometry already
-      // matches these controls.
-      previewSigRef.current = JSON.stringify(controlsRef.current);
-      syncSolving();
-    });
-  }
-
-  useEffect(() => {
-    if (!active) return;
-    const ws = new WebSocket(WS_URL);
-    wsRef.current = ws;
-    // This socket's cuts sender (issue #551). A stable identity per socket
-    // so the close/cleanup handlers only deregister their OWN sender — a
-    // stale socket's late onclose must not tear down the transport a newer
-    // socket just registered.
-    const cutsSender = (msg: string): boolean => {
-      if (ws.readyState !== WebSocket.OPEN) return false;
-      ws.send(msg);
-      return true;
-    };
-    const dropCutsSender = () => {
-      if (cutsWsSend === cutsSender) setCutsWsSend(null);
-      flushCutsWsPending();
-    };
-    ws.onopen = () => {
-      setStatus("open");
-      setCutsWsSend(cutsSender);
-      // A prior socket's in-flight responses can never arrive on this new one.
-      // Treat everything sent so far as received so `solving` can't stick true,
-      // drop stale RTT timers, then send fresh current state. StrictMode and HMR
-      // both tear the socket down + recreate it; the seq counters survive in
-      // refs, so they must never rewind below what's already been received.
-      lastReceivedSeqRef.current = lastSentSeqRef.current;
-      sentAtRef.current.clear();
-      requestSolve();
-    };
-    ws.onclose = () => {
-      setStatus("closed");
-      dropCutsSender();
-      // No solve can progress while disconnected — collapse the outstanding
-      // count so the busy bar can't spin under a "closed" status (reconnect
-      // re-arms it via onopen).
-      lastReceivedSeqRef.current = lastSentSeqRef.current;
-      setSolving(false);
-    };
-    ws.onerror = () => {
-      setStatus("closed");
-      dropCutsSender();
-      lastReceivedSeqRef.current = lastSentSeqRef.current;
-      setSolving(false);
-    };
-    ws.onmessage = (ev) => {
-      const data: SolveResponse & Partial<CutsWsMessage> = JSON.parse(ev.data);
-      if (data._kind === "cuts") {
-        // Cuts sidecar response (issue #551) — never a solve; route it
-        // before any _seq/solving bookkeeping.
-        resolveCutsWsMessage(data as CutsWsMessage);
-        return;
-      }
-      const seq = data._seq ?? 0;
-      // One socket delivers in order, and the server may skip-send superseded
-      // results — so a higher `_seq` implicitly acknowledges every lower one.
-      // Ignore a straggler/duplicate at or below the received watermark.
-      if (seq <= lastReceivedSeqRef.current) {
-        syncSolving();
-        return;
-      }
-      lastReceivedSeqRef.current = seq;
-      // RTT from this seq's send; prune every acked entry (≤ seq) from the map —
-      // seqs skipped server-side never get their own response, so a single
-      // higher-seq arrival clears the whole run of them.
-      const sentAt = sentAtRef.current;
-      const t0 = sentAt.get(seq);
-      if (t0 !== undefined) setRttMs(performance.now() - t0);
-      for (const k of sentAt.keys()) {
-        if (k <= seq) sentAt.delete(k);
-      }
-      // Cancelled through this seq: the user bailed on it (and everything
-      // before). The watermark advanced above so `solving` can clear; just drop
-      // the result rather than rendering it.
-      if (seq <= canceledThroughSeqRef.current) {
-        syncSolving();
-        return;
-      }
-      // Drop a response for an antenna the user already switched away from: a
-      // slow in-flight solve for the previous selection must not stomp the new
-      // antenna's geometry preview (and briefly show the wrong antenna).
-      const staleGeom = !!data.geometry && data.geometry !== geometryRef.current;
-      if (!staleGeom) {
-        if (data.error) {
-          // A solve that raised (e.g. a user design's build_wires) — show the
-          // message and clear stale plot data rather than rendering an empty
-          // result on top of the last antenna.
-          setSolveError(data.error);
-          setResult(null);
-        } else {
-          setSolveError(null);
-          setResult(data);
-        }
-      }
-      syncSolving();
-    };
-    return () => {
-      if (solveRafRef.current !== null) {
-        cancelAnimationFrame(solveRafRef.current);
-        solveRafRef.current = null;
-      }
-      dropCutsSender(); // ws.close() fires onclose async; don't leave a dead sender up
-      ws.close();
-    };
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, [active]);
 
   // Hoisted JSX shared between the desktop tree below and the mobile tree
   // (Phase B). These close over the session's locals, so they are consts /
